auth_service/requirements.txt
passlib[bcrypt]
fastapi
python-multipart
uvicorn
asyncpg
SQLAlchemy
redis
httpx[http2]
pydantic-settings
email-validator==2.3.0
orjson
werkzeug
alembic
backoff==2.2.1
python-jose[cryptography]
asgi-lifespan==2.1.0
gunicorn
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
fastapi-sso
slowapi
